+++ test/Sentry.Tests/Internals/MainSentryEventProcessorTests.cs
@@ -38,6 +38,22 @@ public class MainSentryEventProcessorTests
         Assert.Equal(Environment.UserName, evt.User.Username);
     }
 
+    [Fact]
+    public void EnsureThreadPoolInfoExists()
+    {
+        var evt = new SentryEvent();
+
+        _fixture.SentryOptions.SendDefaultPii = true;
+        var sut = _fixture.GetSut();
+
+        _ = sut.Process(evt);
+        var info = (ThreadPoolInfo)evt.Contexts[MainSentryEventProcessor.ThreadPoolInfoKey];
+        Assert.NotEqual(0, info.MinWorkerThreads);
+        Assert.NotEqual(0, info.MinCompletionPortThreads);
+        Assert.NotEqual(0, info.MaxWorkerThreads);
+        Assert.NotEqual(0, info.MaxCompletionPortThreads);
+    }
+
 #if NETCOREAPP3_1_OR_GREATER
     [Fact]
     public void EnsureMemoryInfoExists()
@@ -50,17 +66,8 @@ public class MainSentryEventProcessorTests
         _ = sut.Process(evt);
         var memory = (MemoryInfo)evt.Contexts[MainSentryEventProcessor.MemoryInfoKey];
         Assert.NotEqual(0, memory.TotalAvailableMemoryBytes);
-        Assert.NotEqual(0, memory.FragmentedBytes);
-        Assert.NotEqual(0, memory.HeapSizeBytes);
         Assert.NotEqual(0, memory.HighMemoryLoadThresholdBytes);
         Assert.NotEqual(0, memory.TotalAvailableMemoryBytes);
-        Assert.NotEqual(0, memory.MemoryLoadBytes);
-#if NET5_0_OR_GREATER
-        Assert.NotEqual(0, memory.TotalCommittedBytes);
-        Assert.NotEqual(0, memory.PromotedBytes);
-        Assert.NotEqual(0, memory.PauseTimePercentage);
-        Assert.NotEmpty(memory.PauseDurations);
-#endif
     }
 #endif