+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Changelog
 
+## 3.13.0
+
+### Features
+
+- Add CaptureLastError as an extension method to the Server class on ASP.NET ([#1411](https://github.com/getsentry/sentry-dotnet/pull/1411))
+- Add IsDynamicCode* to events ([#1418](https://github.com/getsentry/sentry-dotnet/pull/1418))
+
+### Fixes
+
+- Dispose of client should only flush ([#1354](https://github.com/getsentry/sentry-dotnet/pull/1354))
+
 ## 3.12.3
 
 ### Fixes
 
 - Events no longer get dropped because of non-serializable contexts or attachments ([#1401](https://github.com/getsentry/sentry-dotnet/pull/1401))
 - Add MemoryInfo to sentry event ([#1337](https://github.com/getsentry/sentry-dotnet/pull/1337))
+- Report ThreadPool stats ([#1399](https://github.com/getsentry/sentry-dotnet/pull/1399))
 
 ## 3.12.2